server/requirements.txt
einops==0.8.2
fastapi==0.132.0
httpx==0.28.1
lancedb==0.29.2
mcp==1.26.0
openai==2.24.0
pydantic==2.12.5
pydantic-settings==2.12.0
PyYAML==6.0.3
python-docx==1.2.0
python-multipart==0.0.22
pypdf==6.7.5
sentence-transformers==5.1.2
uvicorn==0.40.0
Pillow==12.1.1
pytest==9.0.2
pytest-asyncio==1.3.0
